requirements.txt
Flask
pandas
matplotlib
pymongo
python-dotenv
